update penerimaan pakan dan voadip mobile

+++ application/modules/transaksi/views/penerimaan_pakan_mobile/detail_form.php
@@ -111,15 +111,21 @@
 			</div>
 		</form>
 	</div>
-	<div class="col-lg-12 detailed"><hr></div>
-	<div class="col-xs-12 detailed lock_btn_fiskal" data-date="<?php echo substr($data['tiba'], 0, 10); ?>">
-		<form role="form" class="form-horizontal">
-			<div class="col-xs-6 no-padding" style="padding-right: 5px;">
-				<button type="button" class="btn btn-primary col-xs-12" onclick="ppm.change_tab(this)" data-id="<?php echo $data['no_sj']; ?>" data-noreg="<?php echo $data['noreg']; ?>" data-nomor="<?php echo $data['nomor']; ?>" data-edit="edit" data-href="transaksi"><i class="fa fa-edit"></i> Edit</button>
-			</div>
-			<div class="col-xs-6 no-padding" style="padding-left: 5px;">
-				<button type="button" class="btn btn-danger col-xs-12" onclick="ppm.delete()"><i class="fa fa-trash"></i> Hapus</button>
-			</div>
-		</form>
-	</div>
+	<?php if ($akses['a_submit'] == 1) { ?>
+		<div class="col-lg-12 detailed"><hr></div>
+		<div class="col-xs-12 detailed lock_btn_fiskal" data-date="<?php echo substr($data['tiba'], 0, 10); ?>">
+			<form role="form" class="form-horizontal">
+				<div class="col-xs-6 no-padding" style="padding-right: 5px;">
+					<?php if ($akses['a_edit'] == 1) { ?>
+						<button type="button" class="btn btn-primary col-xs-12" onclick="ppm.change_tab(this)" data-id="<?php echo $data['no_sj']; ?>" data-noreg="<?php echo $data['noreg']; ?>" data-nomor="<?php echo $data['nomor']; ?>" data-edit="edit" data-href="transaksi"><i class="fa fa-edit"></i> Edit</button>
+					<?php } ?>
+				</div>
+				<div class="col-xs-6 no-padding" style="padding-left: 5px;">
+					<?php if ($akses['a_delete'] == 1) { ?>
+						<button type="button" class="btn btn-danger col-xs-12" onclick="ppm.delete()"><i class="fa fa-trash"></i> Hapus</button>
+					<?php } ?>
+				</div>
+			</form>
+		</div>		
+	<?php } ?>
 </div>
